backend/requirements.txt
fastapi
uvicorn
numpy<2.0
scikit-learn==1.2.2
firebase-admin>=6.0.0
python-dotenv
